// src/enums.tsx
export enum findData {
  age = "data-age",
  name = `name="`,
  dataId = `data-id="`,
  dataPosition = "data-positions=",
  title = "<title>",
}

export enum mode {
  filter = "Filter",
  mark = "Mark",
}

export enum pages {
  player = "player",
  team = "team",
  squad = "squad",
  pageSquad = "/squad/",
  all = "All",
}